figma-transformer: extract paint normalizer
